+++ NeuralNetwork.NET/Networks/Implementations/Misc/LayerGradient.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
-using NeuralNetworkNET.Helpers;
 
 namespace NeuralNetworkNET.Networks.Implementations.Misc
 {
@@ -14,14 +13,14 @@ namespace NeuralNetworkNET.Networks.Implementations.Misc
         /// <summary>
         /// Gets the gradient with respect to the current weights
         /// </summary>
-        public double[,] DJdw { get; }
+        public float[,] DJdw { get; }
 
         /// <summary>
         /// Gets the gradient with respect to the current biases
         /// </summary>
-        public double[] Djdb { get; }
+        public float[] Djdb { get; }
 
-        public LayerGradient([NotNull] double[,] dJdw, [NotNull] double[] dJdb)
+        public LayerGradient([NotNull] float[,] dJdw, [NotNull] float[] dJdb)
         {
             DJdw = dJdw;
             Djdb = dJdb;
@@ -40,17 +39,17 @@ namespace NeuralNetworkNET.Networks.Implementations.Misc
         [PublicAPI]
         [Pure, NotNull]
         [CollectionAccess(CollectionAccessType.Read)]
-        public static double[] Flatten([NotNull] this IReadOnlyList<LayerGradient> gradient)
+        public static float[] Flatten([NotNull] this IReadOnlyList<LayerGradient> gradient)
         {
-            if (gradient.Count == 0) return new double[0];
-            double[] result = new double[gradient.Sum(g => g.DJdw.Length + g.Djdb.Length)];
+            if (gradient.Count == 0) return new float[0];
+            float[] result = new float[gradient.Sum(g => g.DJdw.Length + g.Djdb.Length)];
             int position = 0;
             foreach (LayerGradient layer in gradient)
             {
-                int n = sizeof(double) * layer.DJdw.Length;
+                int n = sizeof(float) * layer.DJdw.Length;
                 Buffer.BlockCopy(layer.DJdw, 0, result, position, n);
                 position += n;
-                n = sizeof(double) * layer.Djdb.Length;
+                n = sizeof(float) * layer.Djdb.Length;
                 Buffer.BlockCopy(layer.Djdb, 0, result, position, n);
                 position += n;
             }
