+++ NeuralNetwork.NET/SupervisedLearning/Optimization/Parameters/DatasetParametersBase.cs
@@ -1,15 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace NeuralNetworkNET.SupervisedLearning.Optimization.Parameters
 {
+    /// <summary>
+    /// A base class for an optional dataset to use in a training session
+    /// </summary>
     public abstract class DatasetParametersBase
     {
-        // TODO: add docs
-        public (double[,] X, double[,] Y) Dataset { get; }
+        /// <summary>
+        /// Gets the current dataset
+        /// </summary>
+        public (float[,] X, float[,] Y) Dataset { get; }
 
-        protected DatasetParametersBase((double[,] X, double[,] Y) dataset)
+        protected DatasetParametersBase((float[,] X, float[,] Y) dataset)
         {
             if (dataset.X.GetLength(0) != dataset.Y.GetLength(0)) throw new ArgumentException(nameof(dataset), "The size of the input matrices isn't valid");
             Dataset = dataset;
